requirements.txt
easydict==1.10
matplotlib==3.5.0
numpy==1.24.2
python-dotenv==1.0.0
tensorflow_datasets==4.8.2
tensorflow_gpu==2.9.0
